Avoid integer division.

+++ Character/Class/Warrior/Spells/Rend.cpp
@@ -65,7 +65,7 @@ void Rend::perform_periodic() {
     double damage_dealt = round(damage_remaining / num_ticks_left);
     damage_remaining -= damage_dealt;
 
-    double rage_per_tick = resource_cost / max_ticks;
+    double rage_per_tick = double(resource_cost) / max_ticks;
     double execution_time_per_tick = max_ticks / warr->global_cooldown();
     add_hit_dmg(static_cast<int>(damage_dealt), static_cast<int>(round(rage_per_tick)), execution_time_per_tick);
